+++ core/database.py
@@ -32,14 +32,18 @@ def initialize(database):
     cursor = conn.cursor()
     cursor.execute(
         "CREATE TABLE IF NOT EXISTS 'messages' ('message_id' INT, 'channel' INT,"
-        " 'reactionrole_id' INT, 'guild_id' INT);"
+        " 'reactionrole_id' INT, 'guild_id' INT, 'limit_to_one' INT);"
     )
     cursor.execute(
         "CREATE TABLE IF NOT EXISTS 'reactionroles' ('reactionrole_id' INT, 'reaction'"
         " NVCARCHAR, 'role_id' INT);"
     )
-    cursor.execute("CREATE TABLE IF NOT EXISTS 'admins' ('role_id' INT, 'guild_id' INT);")
-    cursor.execute("CREATE TABLE IF NOT EXISTS 'cleanup_queue_guilds' ('guild_id' INT, 'unix_timestamp' INT);")
+    cursor.execute(
+        "CREATE TABLE IF NOT EXISTS 'admins' ('role_id' INT, 'guild_id' INT);"
+    )
+    cursor.execute(
+        "CREATE TABLE IF NOT EXISTS 'cleanup_queue_guilds' ('guild_id' INT, 'unix_timestamp' INT);"
+    )
     cursor.execute("CREATE TABLE IF NOT EXISTS 'dbinfo' ('version' INT);")
     cursor.execute(
         "CREATE TABLE IF NOT EXISTS 'guild_settings' ('guild_id' INT, 'notify' INT, 'systemchannel' INT);"
@@ -58,6 +62,10 @@ def initialize(database):
     conn.close()
 
 
+class DuplicateInstance(Exception):
+    pass
+
+
 class Database:
     def __init__(self, database):
         self.database = database
@@ -70,26 +78,38 @@ class Database:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
             if self.exists(rl_dict["message"]["message_id"]):
-                raise Exception("The message id is already in use!")
+                raise DuplicateInstance("The message id is already in use!")
             while True:
                 try:
                     reactionrole_id = randint(0, 100000)
                     cursor.execute(
                         "INSERT INTO 'messages' ('message_id', 'channel', 'reactionrole_id',"
-                        " 'guild_id') values(?, ?, ?, ?);",
-                        (rl_dict["message"]["message_id"], rl_dict["message"]["channel_id"], reactionrole_id, rl_dict["message"]["guild_id"]),
+                        " 'guild_id', 'limit_to_one') values(?, ?, ?, ?, ?);",
+                        (
+                            rl_dict["message"]["message_id"],
+                            rl_dict["message"]["channel_id"],
+                            reactionrole_id,
+                            rl_dict["message"]["guild_id"],
+                            rl_dict["limit_to_one"],
+                        ),
                     )
                     break
                 except sqlite3.IntegrityError:
                     continue
-            combos = [(reactionrole_id, reaction, role_id) for reaction, role_id in rl_dict["reactions"].items()]
-            cursor.executemany("INSERT INTO 'reactionroles' ('reactionrole_id', 'reaction', 'role_id') values(?, ?, ?);", combos)
+            combos = [
+                (reactionrole_id, reaction, role_id)
+                for reaction, role_id in rl_dict["reactions"].items()
+            ]
+            cursor.executemany(
+                "INSERT INTO 'reactionroles' ('reactionrole_id', 'reaction', 'role_id') values(?, ?, ?);",
+                combos,
+            )
             conn.commit()
             cursor.close()
             conn.close()
         except sqlite3.Error as e:
             return e
-    
+
     def exists(self, message_id):
         try:
             conn = sqlite3.connect(self.database)
@@ -129,6 +149,21 @@ class Database:
             conn.close()
             return combos
 
+        except sqlite3.Error as e:
+            return e
+
+    def isunique(self, message_id):
+        try:
+            conn = sqlite3.connect(self.database)
+            cursor = conn.cursor()
+            cursor.execute(
+                "SELECT limit_to_one FROM messages WHERE message_id = ?;",
+                (message_id,),
+            )
+            unique = cursor.fetchall()[0][0]
+            cursor.close()
+            conn.close()
+            return unique
         except sqlite3.Error as e:
             return e
 
@@ -212,7 +247,9 @@ class Database:
                 (guild_id,),
             )
             # Delete the guilds potencial cleanup_queue entries
-            cursor.execute("DELETE FROM cleanup_queue_guilds WHERE guild_id=?;", (guild_id,))
+            cursor.execute(
+                "DELETE FROM cleanup_queue_guilds WHERE guild_id=?;", (guild_id,)
+            )
             conn.commit()
 
             cursor.close()
@@ -220,7 +257,7 @@ class Database:
 
         except sqlite3.Error as e:
             return e
-        
+
     def delete(self, message_id, guild_id=None):
         try:
             conn = sqlite3.connect(self.database)
@@ -260,7 +297,10 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("INSERT INTO 'admins' ('role_id', 'guild_id') values(?,?);", (role_id, guild_id))
+            cursor.execute(
+                "INSERT INTO 'admins' ('role_id', 'guild_id') values(?,?);",
+                (role_id, guild_id),
+            )
             conn.commit()
             cursor.close()
             conn.close()
@@ -272,7 +312,10 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("DELETE FROM admins WHERE role_id = ? AND guild_id = ?;", (role_id, guild_id))
+            cursor.execute(
+                "DELETE FROM admins WHERE role_id = ? AND guild_id = ?;",
+                (role_id, guild_id),
+            )
             conn.commit()
             cursor.close()
             conn.close()
@@ -322,7 +365,7 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            channel_id = 0 # Set to false
+            channel_id = 0  # Set to false
             notify = 0
             cursor.execute(
                 "INSERT OR IGNORE INTO guild_settings ('guild_id', 'notify', 'systemchannel')"
@@ -345,7 +388,8 @@ class Database:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
             cursor.execute(
-                "SELECT systemchannel FROM guild_settings WHERE guild_id = ?;", (guild_id,)
+                "SELECT systemchannel FROM guild_settings WHERE guild_id = ?;",
+                (guild_id,),
             )
             result = cursor.fetchall()
             cursor.close()
@@ -376,7 +420,7 @@ class Database:
             for guild in guilds:
                 if guild[0] is not None:
                     guild_ids.append(guild[0])
-            
+
             cursor.close()
             conn.close()
             return guild_ids
@@ -395,7 +439,7 @@ class Database:
             reactionrole_id = cursor.fetchall()[0][0]
             cursor.execute(
                 "SELECT * FROM reactionroles WHERE reactionrole_id = ? AND reaction = ?;",
-                (reactionrole_id, reaction)
+                (reactionrole_id, reaction),
             )
             exists = cursor.fetchall()
             if exists:
@@ -440,7 +484,10 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("INSERT INTO 'cleanup_queue_guilds' ('guild_id', 'unix_timestamp') values(?,?);", (guild_id, unix_timestamp))
+            cursor.execute(
+                "INSERT INTO 'cleanup_queue_guilds' ('guild_id', 'unix_timestamp') values(?,?);",
+                (guild_id, unix_timestamp),
+            )
             conn.commit()
             cursor.close()
             conn.close()
@@ -453,7 +500,9 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("DELETE FROM cleanup_queue_guilds WHERE guild_id=?;", (guild_id,))
+            cursor.execute(
+                "DELETE FROM cleanup_queue_guilds WHERE guild_id=?;", (guild_id,)
+            )
             conn.commit()
             cursor.close()
             conn.close()
@@ -489,22 +538,33 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("SELECT notify FROM guild_settings WHERE guild_id = ?", (guild_id,))
+            cursor.execute(
+                "SELECT notify FROM guild_settings WHERE guild_id = ?", (guild_id,)
+            )
             results = cursor.fetchall()
             if not results:
                 # If the guild was not in the table because the command was never used before
                 notify = 1
                 systemchannel = 0
-                cursor.execute("INSERT INTO 'guild_settings' ('guild_id', 'notify') values(?, ?, ?);", (guild_id, systemchannel, notify))
+                cursor.execute(
+                    "INSERT INTO 'guild_settings' ('guild_id', 'notify') values(?, ?, ?);",
+                    (guild_id, systemchannel, notify),
+                )
             else:
                 notify = results[0][0]
                 if notify:
                     notify = 0
-                    cursor.execute("UPDATE guild_settings SET notify = ? WHERE guild_id = ?", (notify, guild_id))
+                    cursor.execute(
+                        "UPDATE guild_settings SET notify = ? WHERE guild_id = ?",
+                        (notify, guild_id),
+                    )
 
                 else:
                     notify = 1
-                    cursor.execute("UPDATE guild_settings SET notify = ? WHERE guild_id = ?", (notify, guild_id))
+                    cursor.execute(
+                        "UPDATE guild_settings SET notify = ? WHERE guild_id = ?",
+                        (notify, guild_id),
+                    )
             conn.commit()
             cursor.close()
             conn.close()
@@ -519,13 +579,18 @@ class Database:
         try:
             conn = sqlite3.connect(self.database)
             cursor = conn.cursor()
-            cursor.execute("SELECT notify FROM guild_settings WHERE guild_id = ?", (guild_id,))
+            cursor.execute(
+                "SELECT notify FROM guild_settings WHERE guild_id = ?", (guild_id,)
+            )
             results = cursor.fetchall()
             if not results:
                 # If the guild was not in the table because the command was never used before
                 notify = 0
                 systemchannel = 0
-                cursor.execute("INSERT INTO 'guild_settings' ('guild_id', `systemchannel`, 'notify') values(?, ?, ?);", (guild_id, systemchannel, notify))
+                cursor.execute(
+                    "INSERT INTO 'guild_settings' ('guild_id', `systemchannel`, 'notify') values(?, ?, ?);",
+                    (guild_id, systemchannel, notify),
+                )
             else:
                 notify = results[0][0]
             cursor.close()